refactor: simplified blockchains modules in web3
